Player body in rig (#6)
* VRIK and FinalIK added

Follow the HVR documentation for this. Remember to turn ON Inverse Kinematics in the HVRSettings asset in the Assets folder. The player body poser needs to be bound to something.

* FinalIK added to character model

* Update HVRSettings.asset

* Update Level0.unity

* Backup level 0

* Stripped FinalIK demo files from project

* All the character controller stuff

Literally everything from the GDD and the pitch.

Yep. All of it.

Only thing not done is sorting the eye/glasses sometimes mis-grabbing, and it prioritises the eye.

---------

Co-authored-by: HPetch <60263607+HPetch@users.noreply.github.com>
Co-authored-by: HARRY PETCH <petch@harrypet.ch>