Basic rotatable vault door action added
DOES NOT HAVE SEQUENCING (i.e. door does not open at a specific amount of turn - but that's the easy bit). But the interaction works and feels nicer.

Button added which should trigger the lights.
Only thing left to do is make sure that world scan FX gives off an emissive material and that blockout gets a CyberVisible pass.
Should try to art pass the Break Room while we are at it.