Test player character art in level
Scale default by 11 for most accuracy to character controller.
Metallics are too metallic as I don't think we use cubemaps well. Also, with the artstyle, might be best for the metallics to be much more matte... Plus it may be worth changing the colour of the coat to stand out more on the background - brown is blending in currently.