Art optimisation pass (#18)
* Test player character art in level

Scale default by 11 for most accuracy to character controller.
Metallics are too metallic as I don't think we use cubemaps well. Also, with the artstyle, might be best for the metallics to be much more matte... Plus it may be worth changing the colour of the coat to stand out more on the background - brown is blending in currently.

* Occlusion culling data and static+GPU instancing pass

Has a 5 to 10 FPS increase, but weirdly not in the areas it needs. Security Room is not doing it at all. Will figure out why!

Added Lancelot's first char pass too.

* Reflection probes added to scene

* Version 0.7.4 - performance fixes!

Occlusion culling, reflection probes and static batching and GPU instancing added.

* Continued test player character import

* Merge scenes from main correctly