Bird Legs Fixed
If IK ever has a problem, slightly angle the joints in the editor to flow the way you think they should fold - VRIK will sort it out based on the Editor configuration.

No, I don't really understand the words I've just said.

Not pushing this until we fix the model and stop the hands from breaking.